src/plugins/nonebot_plugin_sentry/pyproject.toml
[tool.poetry]
name = "nonebot-plugin-sentry"
version = "0.4.1"
description = "Push your bot errors to Sentry.io"
license = "MIT"
authors = ["yanyongyu <yyy@nonebot.dev>"]
readme = "README.md"
homepage = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry"
repository = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry"
documentation = "https://github.com/cscs181/QQ-GitHub-Bot/tree/master/src/plugins/nonebot_plugin_sentry#readme"
keywords = ["nonebot", "nonebot2", "sentry"]
packages = [
    { include = "nonebot_plugin_sentry/*.py", from = ".." }
]

[tool.poetry.dependencies]
python = "^3.7"
sentry-sdk = "^1.0.0"
nonebot2 = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
